pyproject.toml
[project]
name = "document-format-correction-agent"
version = "0.1.0"
description = "Document format correction agent with a TypeScript runtime and Python host"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-docx>=1.1.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
gui = [
    "PyQt5>=5.15,<6",
    "PyQtWebEngine>=5.15,<6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
